+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Updated `doob/level` to include Server XP
 * Updated `doob/profile` to say "Doob Global XP/Level/Rank".
 * Made unmute/mute commands have embeds.
+* Fixed Level Messages not working
 
 ## v2.0.8
 * Updated description (on the help command) for